Lab 10 - Modified the angular part to account for the requirement changes for lab 10 (ex. a login page)

+++ Lab7/ECommerceStore/ECommerceStore/src/app/album/shopping-cart/shopping-cart.component.ts
@@ -1,7 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import * as constants from "../shared/constants";
-import {Album} from "../shared/album.model";
-import {AlbumService} from "../shared/album.service";
+import * as constants from "../../shared/constants";
+import {Album} from "../../shared/album.model";
+import { AlbumService } from "../../shared/album.service";
+import { Router } from '@angular/router';
+
+class AlbumInShoppingCart {
+  currentAlbum: Album;
+  TimesInCart: number;
+}
 
 @Component({
   selector: 'app-shopping-cart',
@@ -9,15 +15,15 @@ import {AlbumService} from "../shared/album.service";
   styleUrls: ['./shopping-cart.component.css']
 })
 export class ShoppingCartComponent implements OnInit {
-  albums: Album[];
+  albums: AlbumInShoppingCart[];
   currentPage: any;
 
-  constructor(private albumService: AlbumService) {
+  constructor(private albumService: AlbumService, private router: Router) {
     this.currentPage = 1;
     this.albums = []; // when the shopping cart was empty, I would get an error 'albums is undefined', because
     // getAlbumsFromShopping doesn't receive anything
   }
-
+  
   ngOnInit(): void {
     this.getAlbumsFromShoppingCart();
   }
@@ -25,7 +31,14 @@ export class ShoppingCartComponent implements OnInit {
   getAlbumsFromShoppingCart() {
     this.albumService.getAlbumsFromShoppingCart(this.currentPage, constants.ALBUMS_PER_PAGE)
       .subscribe(
-        response => this.albums = response,
+        response => {
+          if (response[0] === "invalid") {
+            this.router.navigate(['']);
+            return;
+          }
+          this.albums.length = 0; // clear the array
+          response.forEach(album => this.albums.push(JSON.parse(album)));
+        },
         error => console.log(error),
       );
   }
@@ -40,17 +53,20 @@ export class ShoppingCartComponent implements OnInit {
   }
 
   goForwardOnePage() {
+    // TO-DO: check that we don't go over the no. of elements in the shopping cart
     this.currentPage += 1;
     this.getAlbumsFromShoppingCart();
   }
 
   private removeFromCart(ID: number, removeOne: boolean) {
-    this.albumService.removeFromShoppingCart(ID, removeOne)
-      .subscribe(
-        response => console.log(response), // it's always gonna be null, but I'm not sure how to remove it
-        error => console.log(error),
-      );
-    this.getAlbumsFromShoppingCart();
+    if (confirm("Are you sure ? Cannot be undone! forta cfr campionii")) {
+      this.albumService.removeFromShoppingCart(ID, removeOne)
+        .subscribe(
+          response => console.log(response), // it's always gonna be null, but I'm not sure how to remove it
+          error => console.log(error),
+        );
+      this.getAlbumsFromShoppingCart();
+    }
   }
 
   removeOneOccurrenceFromCart(ID: number) {
